backend/lesyeux/requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
cssselect2==0.8.0
defusedxml==0.7.1
dj-database-url==2.3.0
Django==4.2.20
django-paypal==2.1
django-ratelimit==4.1.0
django-redis==5.4.0
django-seed==0.3.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
Faker==37.1.0
gunicorn==23.0.0
html5lib==1.1
idna==3.10
lxml==5.4.0
oscrypto==1.3.0
packaging==25.0
pillow==11.2.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
pyHanko==0.26.0
pyhanko-certvalidator==0.26.8
PyJWT==2.9.0
pypdf==5.4.0
python-bidi==0.6.6
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
qrcode==8.1
redis==5.2.1
reportlab==4.4.0
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
toposort==1.10
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uritools==4.0.3
urllib3==2.4.0
webencodings==0.5.1
whitenoise==6.9.0
xhtml2pdf==0.2.17
